+++ lib/impure/rdstdin.nim
@@ -12,8 +12,14 @@
 ## (e.g. you can navigate with the arrow keys). On Windows ``system.readLine``
 ## is used. This suffices because Windows' console already provides the
 ## wanted functionality.
-
-{.deadCodeElim: on.}  # dce option deprecated
+##
+## **Examples:**
+##
+## .. code-block:: nim
+##   echo readLineFromStdin("Is Nim awesome? (Y/n):")
+##   var userResponse: string
+##   doAssert readLineFromStdin("How are you?:", line = userResponse)
+##   echo userResponse
 
 when defined(Windows):
   proc readLineFromStdin*(prompt: string): TaintedString {.
@@ -66,4 +72,3 @@ else:
       historyAdd(buffer)
     linenoise.free(buffer)
     result = true
-