+++ changelog.md
@@ -14,7 +14,7 @@
 ### Breaking changes in the standard library
 
 - `base64.encode` no longer supports `lineLen` and `newLine`.
-  Use `base64.encodeMIME` instead.
+  Use `base64.encodeMime` instead.
 - `os.splitPath()` behavior synchronized with `os.splitFile()` to return "/"
    as the dir component of "/root_sub_dir" instead of the empty string.
 - `sequtils.zip` now returns a sequence of anonymous tuples i.e. those tuples
@@ -123,6 +123,7 @@ echo f
   retrieving the verified certificate chain of the peer we are connected to
   through an SSL-wrapped `Socket`/`AsyncSocket`.
 - Added `distinctBase` overload for values: `assert 12.MyInt.distinctBase == 12`
+- Added `browsers.openDefaultBrowser` without URL, implements IETF RFC-6694 Section-3.
 - Added `jsconsole.trace`, `jsconsole.table`, `jsconsole.exception` for JavaScript target.
 
 
@@ -152,6 +153,9 @@ echo f
 - `net.newContext` now performs SSL Certificate checking on Linux and OSX.
   Define `nimDisableCertificateValidation` to disable it globally.
 - new syntax for lvalue references: `var b {.byaddr.} = expr` enabled by `import pragmas`
+- new module `std/stackframes`, in particular `setFrameMsg` which enables
+  custom runtime annotation of stackframes, see #13351 for examples. Turn on/off via
+  `--stackTraceMsgs:on/off`
 
 ## Language additions
 
@@ -163,6 +167,8 @@ echo f
 
 - `var a {.foo.}: MyType = expr` now lowers to `foo(a, MyType, expr)` for non builtin pragmas,
   enabling things like lvalue references, see `pragmas.byaddr`
+
+- `macro pragmas` can now be used in type sections.
 
 ## Language changes